+++ src/gui/options/mod-options-menu.ts
@@ -233,28 +233,6 @@ modmanager.gui.OptionsMenu = sc.BaseMenu.extend({
         sc.menu.commitHotkeys()
 
         ig.interact.setBlockDelay(0)
-
-        /* refocus the last element, we need to find it first cuz it might have moved to a diffrent position */
-        const element = this.listBox.rowButtonGroup.getCurrentElement() as sc.ButtonGui | undefined
-        if (!element) return
-
-        let y!: number
-        const x = modmanager.gui.optionsMenu.listBox.rowButtonGroup.elements.findIndex(arr => {
-            y = arr.findIndex(e => {
-                if (typeof element.data !== typeof e.data) return false
-                if (typeof e.data === 'object' && typeof element.data === 'object') {
-                    return (
-                        'description' in e.data! &&
-                        'description' in element.data! &&
-                        e.data.description == element.data.description
-                    )
-                } else {
-                    return e.data == element.data
-                }
-            })
-            return y != -1
-        })
-        modmanager.gui.optionsMenu.listBox.rowButtonGroup.focusCurrentButton(y, x, false, true, false)
     },
 })