+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Left clicking a mod in the Settings tab doesn't crash anymore
+- Fix crash when clicking the "Reset Settings" button in CCModManager's mod setting page
 
 ## [1.0.1] 2024-10-28